refactor(#1459): Extract graph traversal engine from ReBACManager (Phase 8)
Extract core Zanzibar-style permission computation into focused modules:

- graph/traversal.py: PermissionComputer class (1,084 LOC)
  - compute_permission(): Core graph traversal with cycle detection
  - compute_permission_with_explanation(): Audit/explain API helper
  - find_direct_relation_tuple(): 3-check lookup (direct, wildcard, userset)
  - has_direct_relation(): Direct relation wrapper
  - format_path_reason(): Human-readable explanation formatter

- graph/expand.py: ExpandEngine class (189 LOC)
  - expand(): Zanzibar Expand API (find all subjects with permission)
  - Recursive expansion via namespace config (union, intersection, TTU)

ReBACManager delegates via composition:
  self._computer = PermissionComputer(repo, get_namespace, max_depth)
  self._expander = ExpandEngine(repo, get_namespace, max_depth)

rebac_manager.py: 3,826 -> 2,941 LOC (-885 LOC, -23%)
Total reduction from original: 4,599 -> 2,941 LOC (-36%)

269 tests pass, 0 regressions. All failures are pre-existing
(Metastore.set_metadata keyword argument issue).
