+++ src/nexus/services/permissions/rebac_manager.py
@@ -26,10 +26,11 @@ from nexus.core.rebac import (
     DEFAULT_PLAYBOOK_NAMESPACE,
     DEFAULT_SKILL_NAMESPACE,
     DEFAULT_TRAJECTORY_NAMESPACE,
-    WILDCARD_SUBJECT,
     Entity,
     NamespaceConfig,
 )
+from nexus.services.permissions.graph.expand import ExpandEngine
+from nexus.services.permissions.graph.traversal import PermissionComputer
 from nexus.services.permissions.rebac_cache import ReBACPermissionCache
 from nexus.services.permissions.rebac_fast import (
     check_permissions_bulk_with_fallback,
@@ -112,6 +113,10 @@ class ReBACManager:
 
         # Issue #1459: Compose TupleRepository for data access delegation
         self._repo = TupleRepository(engine)
+
+        # Issue #1459 Phase 8: Compose graph traversal and expand engines
+        self._computer = PermissionComputer(self._repo, self.get_namespace, max_depth)
+        self._expander = ExpandEngine(self._repo, self.get_namespace, max_depth)
 
         # Deprecation warning for direct ReBACManager instantiation (Phase 2 Task 2.3)
         # Only warn if instantiated directly (not via subclass inheritance)
@@ -2050,35 +2055,9 @@ class ReBACManager:
     ) -> str:
         """Format a permission path into a human-readable reason.
 
-        Args:
-            subject: Subject entity
-            permission: Permission checked
-            obj: Object entity
-            path: Path dictionary from _compute_permission_with_explanation
-
-        Returns:
-            Human-readable explanation string
+        Delegates to PermissionComputer (Issue #1459 Phase 8).
         """
-        parts = []
-        parts.append(f"{subject} has '{permission}' on {obj}")
-
-        # Extract key information from path
-        if "expanded_to" in path:
-            relations = path["expanded_to"]
-            if relations:
-                parts.append(f"(expanded to relations: {', '.join(relations)})")
-
-        if "direct_relation" in path and path["direct_relation"]:
-            parts.append("via direct relation")
-        elif "tupleToUserset" in path:
-            ttu = path["tupleToUserset"]
-            if "found_parents" in ttu and ttu["found_parents"]:
-                parent = ttu["found_parents"][0]
-                parts.append(f"via parent {parent[0]}:{parent[1]}")
-        elif "union" in path:
-            parts.append("via union of relations")
-
-        return " ".join(parts)
+        return PermissionComputer.format_path_reason(subject, permission, obj, path)
 
     def _compute_permission_with_explanation(
         self,
@@ -2090,222 +2069,10 @@ class ReBACManager:
         paths: list[dict[str, Any]],
         zone_id: str | None = None,
     ) -> bool:
-        """Compute permission with detailed path tracking for explanation.
-
-        This is similar to _compute_permission but tracks all paths explored.
-
-        Args:
-            subject: Subject entity
-            permission: Permission to check
-            obj: Object entity
-            visited: Set of visited nodes to detect cycles
-            depth: Current traversal depth
-            paths: List to accumulate path information
-            zone_id: Optional zone ID for multi-zone isolation
-
-        Returns:
-            True if permission is granted
-        """
-        # Initialize path entry
-        path_entry: dict[str, Any] = {
-            "subject": str(subject),
-            "permission": permission,
-            "object": str(obj),
-            "depth": depth,
-            "granted": False,
-        }
-
-        # Check depth limit
-        if depth > self.max_depth:
-            path_entry["error"] = f"Depth limit exceeded (max={self.max_depth})"
-            paths.append(path_entry)
-            return False
-
-        # Check for cycles
-        visit_key = (
-            subject.entity_type,
-            subject.entity_id,
-            permission,
-            obj.entity_type,
-            obj.entity_id,
+        """Compute permission with path tracking. Delegates to PermissionComputer (Issue #1459)."""
+        return self._computer.compute_permission_with_explanation(
+            subject, permission, obj, visited, depth, paths, zone_id
         )
-        if visit_key in visited:
-            path_entry["error"] = "Cycle detected"
-            paths.append(path_entry)
-            return False
-        visited.add(visit_key)
-
-        # Get namespace config
-        namespace = self.get_namespace(obj.entity_type)
-        if not namespace:
-            # No namespace - check direct relation only
-            tuple_info = self._find_direct_relation_tuple(subject, permission, obj, zone_id=zone_id)
-            direct = tuple_info is not None
-            path_entry["direct_relation"] = direct
-            if tuple_info:
-                path_entry["tuple"] = tuple_info
-            path_entry["granted"] = direct
-            paths.append(path_entry)
-            return direct
-
-        # Check if permission is defined explicitly
-        if namespace.has_permission(permission):
-            usersets = namespace.get_permission_usersets(permission)
-            path_entry["expanded_to"] = usersets
-
-            for userset in usersets:
-                userset_sub_paths: list[dict[str, Any]] = []
-                if self._compute_permission_with_explanation(
-                    subject, userset, obj, visited.copy(), depth + 1, userset_sub_paths, zone_id
-                ):
-                    path_entry["granted"] = True
-                    path_entry["via_userset"] = userset
-                    path_entry["sub_paths"] = userset_sub_paths
-                    paths.append(path_entry)
-                    return True
-
-            paths.append(path_entry)
-            return False
-
-        # Check if permission is defined as a relation (legacy)
-        rel_config = namespace.get_relation_config(permission)
-        if not rel_config:
-            # Not defined in namespace - check direct relation
-            tuple_info = self._find_direct_relation_tuple(subject, permission, obj, zone_id=zone_id)
-            direct = tuple_info is not None
-            path_entry["direct_relation"] = direct
-            if tuple_info:
-                path_entry["tuple"] = tuple_info
-            path_entry["granted"] = direct
-            paths.append(path_entry)
-            return direct
-
-        # Handle union
-        if namespace.has_union(permission):
-            union_relations = namespace.get_union_relations(permission)
-            path_entry["union"] = union_relations
-
-            for rel in union_relations:
-                union_sub_paths: list[dict[str, Any]] = []
-                if self._compute_permission_with_explanation(
-                    subject, rel, obj, visited.copy(), depth + 1, union_sub_paths, zone_id
-                ):
-                    path_entry["granted"] = True
-                    path_entry["via_union_member"] = rel
-                    path_entry["sub_paths"] = union_sub_paths
-                    paths.append(path_entry)
-                    return True
-
-            paths.append(path_entry)
-            return False
-
-        # Handle intersection
-        if namespace.has_intersection(permission):
-            intersection_relations = namespace.get_intersection_relations(permission)
-            path_entry["intersection"] = intersection_relations
-            all_granted = True
-
-            for rel in intersection_relations:
-                intersection_sub_paths: list[dict[str, Any]] = []
-                if not self._compute_permission_with_explanation(
-                    subject, rel, obj, visited.copy(), depth + 1, intersection_sub_paths, zone_id
-                ):
-                    all_granted = False
-                    break
-
-            path_entry["granted"] = all_granted
-            paths.append(path_entry)
-            return all_granted
-
-        # Handle exclusion
-        if namespace.has_exclusion(permission):
-            excluded_rel = namespace.get_exclusion_relation(permission)
-            if excluded_rel:
-                exclusion_sub_paths: list[dict[str, Any]] = []
-                has_excluded = self._compute_permission_with_explanation(
-                    subject,
-                    excluded_rel,
-                    obj,
-                    visited.copy(),
-                    depth + 1,
-                    exclusion_sub_paths,
-                    zone_id,
-                )
-                path_entry["exclusion"] = excluded_rel
-                path_entry["granted"] = not has_excluded
-                paths.append(path_entry)
-                return not has_excluded
-
-            paths.append(path_entry)
-            return False
-
-        # Handle tupleToUserset
-        if namespace.has_tuple_to_userset(permission):
-            ttu = namespace.get_tuple_to_userset(permission)
-            if ttu:
-                tupleset_relation = ttu["tupleset"]
-                computed_userset = ttu["computedUserset"]
-
-                # Pattern 1 (parent-style): Find objects where (obj, tupleset_relation, ?)
-                related_objects = self._find_related_objects(obj, tupleset_relation)
-                # Pattern 2 (group-style): Find subjects where (?, tupleset_relation, obj)
-                related_subjects = self._find_subjects_with_relation(obj, tupleset_relation)
-
-                path_entry["tupleToUserset"] = {
-                    "tupleset": tupleset_relation,
-                    "computedUserset": computed_userset,
-                    "found_parents": [(o.entity_type, o.entity_id) for o in related_objects],
-                    "found_subjects": [(s.entity_type, s.entity_id) for s in related_subjects],
-                }
-
-                # Check parent-style relations
-                for related_obj in related_objects:
-                    ttu_sub_paths: list[dict[str, Any]] = []
-                    if self._compute_permission_with_explanation(
-                        subject,
-                        computed_userset,
-                        related_obj,
-                        visited.copy(),
-                        depth + 1,
-                        ttu_sub_paths,
-                        zone_id,
-                    ):
-                        path_entry["granted"] = True
-                        path_entry["sub_paths"] = ttu_sub_paths
-                        path_entry["pattern"] = "parent"
-                        paths.append(path_entry)
-                        return True
-
-                # Check group-style relations
-                for related_subj in related_subjects:
-                    ttu_sub_paths = []
-                    if self._compute_permission_with_explanation(
-                        subject,
-                        computed_userset,
-                        related_subj,
-                        visited.copy(),
-                        depth + 1,
-                        ttu_sub_paths,
-                        zone_id,
-                    ):
-                        path_entry["granted"] = True
-                        path_entry["sub_paths"] = ttu_sub_paths
-                        path_entry["pattern"] = "group"
-                        paths.append(path_entry)
-                        return True
-
-            paths.append(path_entry)
-            return False
-
-        # Direct relation check
-        tuple_info = self._find_direct_relation_tuple(subject, permission, obj, zone_id=zone_id)
-        direct = tuple_info is not None
-        path_entry["direct_relation"] = direct
-        if tuple_info:
-            path_entry["tuple"] = tuple_info
-        path_entry["granted"] = direct
-        paths.append(path_entry)
-        return direct
 
     def _compute_permission(
         self,
@@ -2317,277 +2084,10 @@ class ReBACManager:
         context: dict[str, Any] | None = None,
         zone_id: str | None = None,
     ) -> bool:
-        """Compute permission via graph traversal.
-
-        Args:
-            subject: Subject entity
-            permission: Permission to check (can be string or userset dict)
-            obj: Object entity
-            visited: Set of visited (subject_type, subject_id, permission, object_type, object_id) to detect cycles
-            depth: Current traversal depth
-            context: Optional ABAC context for condition evaluation
-            zone_id: Optional zone ID for multi-zone isolation
-
-        Returns:
-            True if permission is granted
-        """
-        # P0-6: Explicit deny on graph traversal limit exceeded
-        # Security policy: ALWAYS deny when graph is too deep (never allow)
-        if depth > self.max_depth:
-            logger.debug(
-                f"ReBAC graph traversal depth limit exceeded (max={self.max_depth}): "
-                f"DENYING permission '{permission}' for {subject} -> {obj}"
-            )
-            return False  # EXPLICIT DENY - never allow on limit exceed
-
-        # P0-6: Check for cycles (prevent infinite loops)
-        # Convert permission to hashable string for visit_key
-        permission_key = (
-            json.dumps(permission, sort_keys=True) if isinstance(permission, dict) else permission
+        """Compute permission via graph traversal. Delegates to PermissionComputer (Issue #1459)."""
+        return self._computer.compute_permission(
+            subject, permission, obj, visited, depth, context, zone_id
         )
-        visit_key = (
-            subject.entity_type,
-            subject.entity_id,
-            permission_key,
-            obj.entity_type,
-            obj.entity_id,
-        )
-        if visit_key in visited:
-            # Cycle detected - deny to prevent infinite loop
-            logger.debug(
-                f"ReBAC graph cycle detected: DENYING permission '{permission}' "
-                f"for {subject} -> {obj} (already visited)"
-            )
-            return False  # EXPLICIT DENY - never allow cycles
-        visited.add(visit_key)
-
-        # Handle dict permission (userset rewrite rules from Zanzibar)
-        if isinstance(permission, dict):
-            # Handle "this" - direct relation check
-            if "this" in permission:
-                # Check if there's a direct tuple (any relation works for "this")
-                # In Zanzibar, "this" means the relation itself
-                # This is used when the relation config is like: {"union": [{"this": {}}, ...]}
-                # For now, we treat "this" as checking the relation name from context
-                # Since we don't have the relation name in dict form, skip "this" handling
-                # The caller should pass the relation name as a string, not {"this": {}}
-                return False
-
-            # Handle "computed_userset" - check a specific relation on the same object
-            if "computed_userset" in permission:
-                computed = permission["computed_userset"]
-                if isinstance(computed, dict):
-                    # Extract relation from computed_userset
-                    # Format: {"object": ".", "relation": "viewer"}
-                    # "." means the same object
-                    relation_name = computed.get("relation")
-                    if relation_name:
-                        # Recursively check the relation
-                        return self._compute_permission(
-                            subject,
-                            relation_name,
-                            obj,
-                            visited.copy(),
-                            depth + 1,
-                            context,
-                            zone_id,
-                        )
-                return False
-
-            # Unknown dict format - deny
-            logger.warning(f"Unknown permission dict format: {permission}")
-            return False
-
-        # Get namespace config for object type
-        namespace = self.get_namespace(obj.entity_type)
-        if not namespace:
-            # No namespace config - check for direct relation only
-            logger.debug(
-                f"  [depth={depth}] ⚠️ No namespace for {obj.entity_type}, checking direct relation"
-            )
-            return self._has_direct_relation(subject, permission, obj, context, zone_id)
-
-        logger.debug(f"  [depth={depth}] ✅ Found namespace for {obj.entity_type}")
-        logger.debug(
-            f"  [depth={depth}] 📊 ALL Relations in namespace: {list(namespace.config.get('relations', {}).keys())}"
-        )
-        logger.debug(
-            f"  [depth={depth}] 📊 ALL Permissions in namespace: {list(namespace.config.get('permissions', {}).keys())}"
-        )
-
-        # P0-1: Use explicit permission-to-userset mapping (Zanzibar-style)
-        # Check if permission is defined via "permissions" config (new way)
-        if namespace.has_permission(permission):
-            # Permission defined explicitly - check all usersets that grant it
-            usersets = namespace.get_permission_usersets(permission)
-            logger.debug(
-                f"  [depth={depth}] 🔑 Permission '{permission}' defined in namespace '{obj.entity_type}'"
-            )
-            logger.debug(
-                f"  [depth={depth}] 📋 Permission '{permission}' expands to usersets: {usersets}"
-            )
-            logger.debug(
-                f"  [depth={depth}] 🧪 Checking {len(usersets)} usersets for {subject} on {obj}"
-            )
-            logger.debug(
-                f"  [depth={depth}] 📊 NAMESPACE CONFIG for '{obj.entity_type}': relations={list(namespace.config.get('relations', {}).keys())}"
-            )
-
-            for i, userset in enumerate(usersets):
-                logger.debug(
-                    f"  [depth={depth}] 🔍 [{i + 1}/{len(usersets)}] Checking userset '{userset}'..."
-                )
-                result = self._compute_permission(
-                    subject, userset, obj, visited.copy(), depth + 1, context, zone_id
-                )
-                if result:
-                    logger.debug(
-                        f"  [depth={depth}] ✅ [{i + 1}/{len(usersets)}] GRANTED via userset '{userset}'"
-                    )
-                    return True
-                else:
-                    logger.debug(
-                        f"  [depth={depth}] ❌ [{i + 1}/{len(usersets)}] DENIED for userset '{userset}'"
-                    )
-
-            logger.debug(
-                f"  [depth={depth}] 🚫 ALL {len(usersets)} usersets DENIED - permission DENIED"
-            )
-            return False
-
-        # Fallback: Check if permission is defined as a relation (legacy)
-        rel_config = namespace.get_relation_config(permission)
-        logger.debug(
-            f"  [depth={depth}] 🔍 Checking relation config for '{permission}': {rel_config}"
-        )
-        if not rel_config:
-            # Permission not defined in namespace - check for direct relation
-            logger.debug(
-                f"  [depth={depth}] ⚠️ No relation config for '{permission}', checking direct relation"
-            )
-            return self._has_direct_relation(subject, permission, obj, context, zone_id)
-
-        # Handle union (OR of multiple relations)
-        if namespace.has_union(permission):
-            union_relations = namespace.get_union_relations(permission)
-            logger.debug(
-                f"  [depth={depth}] 🔗 Relation '{permission}' is UNION of: {union_relations}"
-            )
-            logger.debug(
-                f"  [depth={depth}] 📋 Relation config for '{permission}': {namespace.get_relation_config(permission)}"
-            )
-            for i, rel in enumerate(union_relations):
-                logger.debug(
-                    f"  [depth={depth}] 🔍 [{i + 1}/{len(union_relations)}] Checking union relation '{rel}'..."
-                )
-                result = self._compute_permission(
-                    subject, rel, obj, visited.copy(), depth + 1, context, zone_id
-                )
-                if result:
-                    logger.debug(
-                        f"  [depth={depth}] ✅ [{i + 1}/{len(union_relations)}] GRANTED via union relation '{rel}'"
-                    )
-                    return True
-                else:
-                    logger.debug(
-                        f"  [depth={depth}] ❌ [{i + 1}/{len(union_relations)}] DENIED for union relation '{rel}'"
-                    )
-            logger.debug(f"  [depth={depth}] 🚫 ALL union relations DENIED")
-            return False
-
-        # Handle intersection (AND of multiple relations)
-        if namespace.has_intersection(permission):
-            intersection_relations = namespace.get_intersection_relations(permission)
-            # ALL relations must be true
-            for rel in intersection_relations:
-                if not self._compute_permission(
-                    subject, rel, obj, visited.copy(), depth + 1, context, zone_id
-                ):
-                    return False  # If any relation is False, whole intersection is False
-            return True  # All relations were True
-
-        # Handle exclusion (NOT relation - this implements DENY semantics)
-        if namespace.has_exclusion(permission):
-            excluded_rel = namespace.get_exclusion_relation(permission)
-            if excluded_rel:
-                # Must NOT have the excluded relation
-                return not self._compute_permission(
-                    subject, excluded_rel, obj, visited.copy(), depth + 1, context, zone_id
-                )
-            return False
-
-        # Handle tupleToUserset (indirect relation via another object)
-        if namespace.has_tuple_to_userset(permission):
-            ttu = namespace.get_tuple_to_userset(permission)
-            logger.debug(f"  [depth={depth}] 🔄 tupleToUserset for '{permission}': {ttu}")
-            if ttu:
-                tupleset_relation = ttu["tupleset"]
-                computed_userset = ttu["computedUserset"]
-
-                # Pattern 1 (parent-style): Find objects where (obj, tupleset_relation, ?)
-                # Example: (child_file, "parent", parent_dir) -> check subject has computed_userset on parent_dir
-                related_objects = self._find_related_objects(obj, tupleset_relation)
-                logger.debug(
-                    f"  [depth={depth}] 🔍 Pattern 1 (parent): Found {len(related_objects)} related objects via tupleset '{tupleset_relation}': {[(o.entity_type, o.entity_id) for o in related_objects]}"
-                )
-
-                # Check if subject has computed_userset on any related object
-                for i, related_obj in enumerate(related_objects):
-                    logger.debug(
-                        f"  [depth={depth}] 🔍 [{i + 1}/{len(related_objects)}] Checking if {subject} has '{computed_userset}' on {related_obj}..."
-                    )
-                    if self._compute_permission(
-                        subject,
-                        computed_userset,
-                        related_obj,
-                        visited.copy(),
-                        depth + 1,
-                        context,
-                        zone_id,
-                    ):
-                        logger.debug(
-                            f"  [depth={depth}] ✅ GRANTED via tupleToUserset (parent pattern) through {related_obj}"
-                        )
-                        return True
-                    else:
-                        logger.debug(f"  [depth={depth}] ❌ DENIED for {related_obj}")
-
-                # Pattern 2 (group-style): Find subjects where (?, tupleset_relation, obj)
-                # Example: (group, "direct_viewer", file) -> check subject has computed_userset on group
-                related_subjects = self._find_subjects_with_relation(obj, tupleset_relation)
-                logger.debug(
-                    f"  [depth={depth}] 🔍 Pattern 2 (group): Found {len(related_subjects)} subjects with '{tupleset_relation}' on {obj}: {[(s.entity_type, s.entity_id) for s in related_subjects]}"
-                )
-
-                # Check if subject has computed_userset on any related subject (typically group membership)
-                for i, related_subj in enumerate(related_subjects):
-                    logger.debug(
-                        f"  [depth={depth}] 🔍 [{i + 1}/{len(related_subjects)}] Checking if {subject} has '{computed_userset}' on {related_subj}..."
-                    )
-                    if self._compute_permission(
-                        subject,
-                        computed_userset,
-                        related_subj,
-                        visited.copy(),
-                        depth + 1,
-                        context,
-                        zone_id,
-                    ):
-                        logger.debug(
-                            f"  [depth={depth}] ✅ GRANTED via tupleToUserset (group pattern) through {related_subj}"
-                        )
-                        return True
-                    else:
-                        logger.debug(f"  [depth={depth}] ❌ DENIED for {related_subj}")
-
-                logger.debug(
-                    f"  [depth={depth}] 🚫 tupleToUserset: No related objects/subjects granted permission"
-                )
-
-            return False
-
-        # Direct relation check (with optional context evaluation)
-        return self._has_direct_relation(subject, permission, obj, context, zone_id)
 
     def _has_direct_relation(
         self,
@@ -2597,35 +2097,8 @@ class ReBACManager:
         context: dict[str, Any] | None = None,
         zone_id: str | None = None,
     ) -> bool:
-        """Check if subject has direct relation to object.
-
-        Checks both:
-        1. Direct concrete subject tuple: (subject, relation, object)
-        2. Userset-as-subject tuple: (subject_set#set_relation, relation, object)
-           where subject has set_relation on subject_set
-
-        If context is provided, evaluates tuple conditions (ABAC).
-
-        Args:
-            subject: Subject entity
-            relation: Relation type
-            obj: Object entity
-            context: Optional ABAC context for condition evaluation
-            zone_id: Optional zone ID for multi-zone isolation
-
-        Returns:
-            True if direct relation exists and conditions are satisfied
-        """
-        logger.debug(
-            f"    💾 Checking DATABASE for direct tuple: subject={subject}, relation={relation}, object={obj}, zone_id={zone_id}"
-        )
-        result = self._find_direct_relation_tuple(subject, relation, obj, context, zone_id)
-        if result is not None:
-            logger.debug(f"    ✅ FOUND tuple: {result.get('tuple_id', 'unknown')}")
-            return True
-        else:
-            logger.debug("    ❌ NO tuple found in database")
-            return False
+        """Check if subject has direct relation. Delegates to PermissionComputer (Issue #1459)."""
+        return self._computer.has_direct_relation(subject, relation, obj, context, zone_id)
 
     def _find_direct_relation_tuple(
         self,
@@ -2635,235 +2108,10 @@ class ReBACManager:
         context: dict[str, Any] | None = None,
         zone_id: str | None = None,
     ) -> dict[str, Any] | None:
-        """Find direct relation tuple with full details.
-
-        Returns tuple information for explain API.
-
-        Args:
-            subject: Subject entity
-            relation: Relation type
-            obj: Object entity
-            context: Optional ABAC context for condition evaluation
-            zone_id: Optional zone ID for multi-zone isolation
-
-        Returns:
-            Tuple dict with id, subject, relation, object info, or None if not found
-        """
-        logger.debug(
-            f"    _find_direct_relation_tuple: subject={subject}, relation={relation}, obj={obj}, zone_id={zone_id}"
+        """Find direct relation tuple. Delegates to PermissionComputer (Issue #1459)."""
+        return self._computer.find_direct_relation_tuple(
+            subject, relation, obj, context, zone_id
         )
-
-        with self._connection() as conn:
-            cursor = self._create_cursor(conn)
-
-            # BUGFIX: Use >= instead of > for exact expiration boundary
-            # Check 1: Direct concrete subject (subject_relation IS NULL)
-            # ABAC: Fetch conditions column to evaluate context
-            # P0-4: Filter by zone_id for multi-zone isolation
-            if zone_id is None:
-                cursor.execute(
-                    self._fix_sql_placeholders(
-                        """
-                        SELECT tuple_id, subject_type, subject_id, subject_relation,
-                               relation, object_type, object_id, conditions, expires_at
-                        FROM rebac_tuples
-                        WHERE subject_type = ? AND subject_id = ?
-                          AND subject_relation IS NULL
-                          AND relation = ?
-                          AND object_type = ? AND object_id = ?
-                          AND (expires_at IS NULL OR expires_at >= ?)
-                          AND zone_id IS NULL
-                        LIMIT 1
-                        """
-                    ),
-                    (
-                        subject.entity_type,
-                        subject.entity_id,
-                        relation,
-                        obj.entity_type,
-                        obj.entity_id,
-                        datetime.now(UTC).isoformat(),
-                    ),
-                )
-            else:
-                cursor.execute(
-                    self._fix_sql_placeholders(
-                        """
-                        SELECT tuple_id, subject_type, subject_id, subject_relation,
-                               relation, object_type, object_id, conditions, expires_at
-                        FROM rebac_tuples
-                        WHERE subject_type = ? AND subject_id = ?
-                          AND subject_relation IS NULL
-                          AND relation = ?
-                          AND object_type = ? AND object_id = ?
-                          AND (expires_at IS NULL OR expires_at >= ?)
-                          AND zone_id = ?
-                        LIMIT 1
-                        """
-                    ),
-                    (
-                        subject.entity_type,
-                        subject.entity_id,
-                        relation,
-                        obj.entity_type,
-                        obj.entity_id,
-                        datetime.now(UTC).isoformat(),
-                        zone_id,
-                    ),
-                )
-
-            row = cursor.fetchone()
-            if row:
-                logger.debug(f"    ✅ Found direct tuple for {subject} -> {relation} -> {obj}")
-                # Tuple exists - now check conditions if context provided
-                conditions_json = row["conditions"]
-
-                if conditions_json:
-                    try:
-                        conditions = (
-                            json.loads(conditions_json)
-                            if isinstance(conditions_json, str)
-                            else conditions_json
-                        )
-                        # Evaluate ABAC conditions
-                        if not self._evaluate_conditions(conditions, context):
-                            logger.debug(
-                                f"Tuple exists but conditions not satisfied for {subject} -> {relation} -> {obj}"
-                            )
-                            return None  # Tuple exists but conditions failed
-                    except (json.JSONDecodeError, TypeError) as e:
-                        logger.warning(f"Failed to parse conditions JSON: {e}")
-                        # On parse error, treat as no conditions (allow)
-
-                # Return tuple details
-                return dict(row)
-            else:
-                logger.debug(f"    ❌ No direct tuple found for {subject} -> {relation} -> {obj}")
-
-            # Check 2: Wildcard/public access
-            # Check if wildcard subject (*:*) has the relation (public access)
-            # Avoid infinite recursion by only checking wildcard if subject is not already wildcard
-            # P0-4: Filter by zone_id for multi-zone isolation
-            if (subject.entity_type, subject.entity_id) != WILDCARD_SUBJECT:
-                wildcard_entity = Entity(WILDCARD_SUBJECT[0], WILDCARD_SUBJECT[1])
-                if zone_id is None:
-                    cursor.execute(
-                        self._fix_sql_placeholders(
-                            """
-                            SELECT tuple_id, subject_type, subject_id, subject_relation,
-                                   relation, object_type, object_id, conditions, expires_at
-                            FROM rebac_tuples
-                            WHERE subject_type = ? AND subject_id = ?
-                              AND subject_relation IS NULL
-                              AND relation = ?
-                              AND object_type = ? AND object_id = ?
-                              AND (expires_at IS NULL OR expires_at >= ?)
-                              AND zone_id IS NULL
-                            LIMIT 1
-                            """
-                        ),
-                        (
-                            wildcard_entity.entity_type,
-                            wildcard_entity.entity_id,
-                            relation,
-                            obj.entity_type,
-                            obj.entity_id,
-                            datetime.now(UTC).isoformat(),
-                        ),
-                    )
-                else:
-                    cursor.execute(
-                        self._fix_sql_placeholders(
-                            """
-                            SELECT tuple_id, subject_type, subject_id, subject_relation,
-                                   relation, object_type, object_id, conditions, expires_at
-                            FROM rebac_tuples
-                            WHERE subject_type = ? AND subject_id = ?
-                              AND subject_relation IS NULL
-                              AND relation = ?
-                              AND object_type = ? AND object_id = ?
-                              AND (expires_at IS NULL OR expires_at >= ?)
-                              AND zone_id = ?
-                            LIMIT 1
-                            """
-                        ),
-                        (
-                            wildcard_entity.entity_type,
-                            wildcard_entity.entity_id,
-                            relation,
-                            obj.entity_type,
-                            obj.entity_id,
-                            datetime.now(UTC).isoformat(),
-                            zone_id,
-                        ),
-                    )
-                row = cursor.fetchone()
-                if row:
-                    return dict(row)
-
-                # Check 2b: Cross-zone wildcard access (Issue #1064)
-                # Wildcards should grant access across ALL zones, not just the owner's zone.
-                # This is the industry-standard pattern used by SpiceDB, OpenFGA, and Ory Keto.
-                # Only check cross-zone if zone_id is provided (multi-zone mode).
-                if zone_id is not None:
-                    cursor.execute(
-                        self._fix_sql_placeholders(
-                            """
-                            SELECT tuple_id, subject_type, subject_id, subject_relation,
-                                   relation, object_type, object_id, conditions, expires_at
-                            FROM rebac_tuples
-                            WHERE subject_type = ? AND subject_id = ?
-                              AND subject_relation IS NULL
-                              AND relation = ?
-                              AND object_type = ? AND object_id = ?
-                              AND (expires_at IS NULL OR expires_at >= ?)
-                            LIMIT 1
-                            """
-                        ),
-                        (
-                            wildcard_entity.entity_type,
-                            wildcard_entity.entity_id,
-                            relation,
-                            obj.entity_type,
-                            obj.entity_id,
-                            datetime.now(UTC).isoformat(),
-                        ),
-                    )
-                    row = cursor.fetchone()
-                    if row:
-                        logger.debug(f"    Cross-zone wildcard access: *:* -> {relation} -> {obj}")
-                        return dict(row)
-
-            # Check 3: Userset-as-subject grants
-            # Find tuples like (group:eng#member, editor-of, file:readme)
-            # where subject has 'member' relation to 'group:eng'
-            subject_sets = self._find_subject_sets(relation, obj, zone_id)
-            for set_type, set_id, set_relation in subject_sets:
-                # Recursively check if subject has set_relation on the set entity
-                if self._has_direct_relation(
-                    subject, set_relation, Entity(set_type, set_id), context, zone_id
-                ):
-                    # Return the userset tuple that granted access
-                    cursor.execute(
-                        self._fix_sql_placeholders(
-                            """
-                            SELECT tuple_id, subject_type, subject_id, subject_relation,
-                                   relation, object_type, object_id, conditions, expires_at
-                            FROM rebac_tuples
-                            WHERE subject_type = ? AND subject_id = ?
-                              AND subject_relation = ?
-                              AND relation = ?
-                              AND object_type = ? AND object_id = ?
-                            LIMIT 1
-                            """
-                        ),
-                        (set_type, set_id, set_relation, relation, obj.entity_type, obj.entity_id),
-                    )
-                    row = cursor.fetchone()
-                    if row:
-                        return dict(row)
-
-            return None
 
     def _find_subject_sets(
         self, relation: str, obj: Entity, zone_id: str | None = None
@@ -2892,142 +2140,9 @@ class ReBACManager:
     ) -> list[tuple[str, str]]:
         """Find all subjects with a given permission on an object.
 
-        Args:
-            permission: Permission to check
-            object: (object_type, object_id) tuple
-
-        Returns:
-            List of (subject_type, subject_id) tuples
-
-        Example:
-            >>> manager.rebac_expand(
-            ...     permission="read",
-            ...     object=("file", "file_id")
-            ... )
-            [("agent", "alice_id"), ("agent", "bob_id")]
+        Delegates to ExpandEngine (Issue #1459 Phase 8).
         """
-        object_entity = Entity(object[0], object[1])
-        subjects: set[tuple[str, str]] = set()
-
-        # Get namespace config
-        namespace = self.get_namespace(object_entity.entity_type)
-        if not namespace:
-            # No namespace - return direct relations only
-            return self._get_direct_subjects(permission, object_entity)
-
-        # Recursively expand permission via namespace config
-        self._expand_permission(
-            permission, object_entity, namespace, subjects, visited=set(), depth=0
-        )
-
-        return list(subjects)
-
-    def _expand_permission(
-        self,
-        permission: str,
-        obj: Entity,
-        namespace: NamespaceConfig,
-        subjects: set[tuple[str, str]],
-        visited: set[tuple[str, str, str]],
-        depth: int,
-    ) -> None:
-        """Recursively expand permission to find all subjects.
-
-        Args:
-            permission: Permission to expand
-            obj: Object entity
-            namespace: Namespace configuration
-            subjects: Set to accumulate subjects
-            visited: Set of visited (permission, object_type, object_id) to detect cycles
-            depth: Current traversal depth
-        """
-        # Check depth limit
-        if depth > self.max_depth:
-            return
-
-        # Check for cycles
-        visit_key = (permission, obj.entity_type, obj.entity_id)
-        if visit_key in visited:
-            return
-        visited.add(visit_key)
-
-        # Get relation config
-        rel_config = namespace.get_relation_config(permission)
-        if not rel_config:
-            # Permission not defined in namespace - check for direct relations
-            direct_subjects = self._get_direct_subjects(permission, obj)
-            for subj in direct_subjects:
-                subjects.add(subj)
-            return
-
-        # Handle union
-        if namespace.has_union(permission):
-            union_relations = namespace.get_union_relations(permission)
-            for rel in union_relations:
-                self._expand_permission(rel, obj, namespace, subjects, visited.copy(), depth + 1)
-            return
-
-        # Handle intersection (find subjects that have ALL relations)
-        if namespace.has_intersection(permission):
-            intersection_relations = namespace.get_intersection_relations(permission)
-            if not intersection_relations:
-                return
-
-            # Get subjects for each relation
-            relation_subjects = []
-            for rel in intersection_relations:
-                rel_subjects: set[tuple[str, str]] = set()
-                self._expand_permission(
-                    rel, obj, namespace, rel_subjects, visited.copy(), depth + 1
-                )
-                relation_subjects.append(rel_subjects)
-
-            # Find intersection (subjects that appear in ALL sets)
-            if relation_subjects:
-                common_subjects = set.intersection(*relation_subjects)
-                for subj in common_subjects:
-                    subjects.add(subj)
-            return
-
-        # Handle exclusion (find subjects that DON'T have the excluded relation)
-        if namespace.has_exclusion(permission):
-            # Note: Expand for exclusion is complex and potentially expensive
-            # We would need to find all possible subjects, then filter out those with the excluded relation
-            # For now, we skip expand for exclusion relations
-            # TODO: Implement if needed for production use
-            logger.warning(
-                f"Expand API does not support exclusion relations yet: {permission} on {obj}"
-            )
-            return
-
-        # Handle tupleToUserset
-        if namespace.has_tuple_to_userset(permission):
-            ttu = namespace.get_tuple_to_userset(permission)
-            if ttu:
-                tupleset_relation = ttu["tupleset"]
-                computed_userset = ttu["computedUserset"]
-
-                # Find all related objects
-                related_objects = self._find_related_objects(obj, tupleset_relation)
-
-                # Expand permission on related objects
-                for related_obj in related_objects:
-                    related_ns = self.get_namespace(related_obj.entity_type)
-                    if related_ns:
-                        self._expand_permission(
-                            computed_userset,
-                            related_obj,
-                            related_ns,
-                            subjects,
-                            visited.copy(),
-                            depth + 1,
-                        )
-            return
-
-        # Direct relation - add all subjects
-        direct_subjects = self._get_direct_subjects(permission, obj)
-        for subj in direct_subjects:
-            subjects.add(subj)
+        return self._expander.expand(permission, object)
 
     def _get_direct_subjects(self, relation: str, obj: Entity) -> list[tuple[str, str]]:
         """Get all subjects with direct relation to object. Delegates to TupleRepository."""
